requirements/base.txt
django==3.2
django-filter==2.4.0
dj-database-url==0.5.0
psycopg2-binary==2.8.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
